feat: implement toggle completion functionality and enhance layout in ItemDetailsPage and List components
