+++ src/components/ItemDetailsPage.jsx
@@ -1,19 +1,19 @@
 import React from "react";
-import { useParams } from "react-router-dom";
+import { Link, useParams } from "react-router-dom";
 import yes from "../assets/check_mark.png";
 import no from "../assets/x.png"; 
 import {useState} from "react";
 
 import styles from "./ItemDetailsPage.module.css";
 
-export default function ItemDetailsPage({dataFromMyParent}) {
+export default function ItemDetailsPage({dataFromMyParent, toggleCompletion}) {
   const { id } = useParams();
   const item = dataFromMyParent.find((item) => item.id === id);
-  const [completed, setCompleted] = useState(item.completed)
 
-  const trueFalseToggle = () => {
-    setCompleted(prevCompleted => !prevCompleted)
-  };
+
+  // const trueFalseToggle = () => {
+  //   setCompleted(prevCompleted => !prevCompleted)
+  // };
 
   if (!item) {
     return <div>Item not found</div>;
@@ -23,21 +23,24 @@ export default function ItemDetailsPage({dataFromMyParent}) {
     <>
       <div className={styles.itemDetails}>
         <div>
-          <h1>Item Detailssss</h1>
+          <h1>Item Details</h1>
           <p>
             <strong>Task:</strong> {item.task}
           </p>
           <p>
-            <strong>Completed:</strong> {completed ? "Yes" : "No"}
+            <strong>Completed:</strong> {item.completed ? "Yes" : "No"}
           </p>
         </div>
-        <div>
+        <div className={styles.imgAndBack}>
           <img
-            src={completed ? yes : no}
-            alt={completed ? "Completed" : "Not Completed"}
-            onClick={trueFalseToggle}
+            src={item.completed ? yes : no}
+            alt={item.completed ? "Completed" : "Not Completed"}
+            // onClick={trueFalseToggle}
             onClick={() => toggleCompletion(item.id)}
           />
+          <Link to="/">
+          <button>Back</button>
+          </Link>
         </div>
       </div>
     </>
