+++ src/components/ItemDetailsPage.module.css
@@ -18,3 +18,21 @@
     font-size: 1.2em;
     margin-bottom: 10px;
 }
+
+.imgAndBack {
+    display: flex;
+    flex-direction: column;
+    align-items: end;
+    justify-content: space-between;
+}
+
+.imgAndBack button {
+    background-color: rgb(153, 135, 252);
+    color: white;
+    border: none;
+    padding: 0.5em 1em;
+    border-radius: 5px;
+    margin-left: 0.5em;
+    margin-right: 0;
+    cursor: pointer;
+}
